requirements.txt
langchain>=0.2.8
langchain-community>=0.2.7
pydantic>=2.6
python-dotenv>=1.0
pypdf>=4.2
pdfplumber>=0.11
pinecone>=3.0
pytest>=8.2